+++ CHANGELOG.md
@@ -4,7 +4,6 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### 0.0.2 (2020-12-02)
 
-
 ### Features
 
-* 🎸 add validator components ([d9d4b71](https://github.com/21epub/epub-form-validation/commit/d9d4b71828246f96e8e246d14237bc149e07b54c))
+- 🎸 add validator components ([d9d4b71](https://github.com/21epub/epub-form-validation/commit/d9d4b71828246f96e8e246d14237bc149e07b54c))